refactor (YH-25) findUserById query operation
